+++ app/javascript/bundles/Container/components/lessons/AlkaliMetals.js
@@ -1,12 +1,18 @@
 import React from "react";
 import Text from "./text/Text";
-// import AlkaliMetalsTHREE from "./AlkaliMetalsTHREE";
+import AlkaliMetalsThree from "./AlkaliMetalsThree";
 
-const AlkaliMetals = (props) => (
+const AlkaliMetals = props => {
+  
+  return (
     <div>
+      <section id="container" className="lesson-container">
         
-        <Text props={props.description} />
+        <AlkaliMetalsThree />
+      </section>
+      <Text props={props.description} />
     </div>
-);
+  );
+};
 
 export default props => <AlkaliMetals {...props} />;
