+++ app/javascript/scimo/scripts/three.js
@@ -1,3 +1,6 @@
+import "./tween.min.js";
+import "./TrackballControls.js";
+
 (function (global, factory) {
 	typeof exports === 'object' && typeof module !== 'undefined' ? factory(exports) :
 	typeof define === 'function' && define.amd ? define(['exports'], factory) :
