+++ app/javascript/scimo/styles/AlkaliMetals2.css
@@ -1,4 +1,4 @@
-html, body {
+/* html, body {
     height: 100%;
 }
 body {
@@ -65,4 +65,4 @@ button:hover {
 button:active {
     color: #000000;
     background-color: rgba(0,255,255,0.75);
-}
+} */
